pyproject.toml
[tool.poetry]
name = "bootcamp-ultimo-dia"
version = "0.1.0"
description = ""
authors = ["lvgalvaofilho"]
readme = "README.md"

[tool.poetry.dependencies]
python = "==3.11.5"
dbt-postgres = "^1.7.12"
apache-airflow = "^2.9.0"
apache-airflow-providers-dbt-cloud = "^3.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
